+++ device-noolite/src/main/java/ru/iris/noolite/protocol/events/NooliteUnbindRXChannel.java
@@ -4,17 +4,17 @@ import ru.iris.commons.bus.AbstractEvent;
 
 public class NooliteUnbindRXChannel extends AbstractEvent {
 
-	private byte channel;
+	private Short channel;
 
-	public NooliteUnbindRXChannel(byte channel) {
+	public NooliteUnbindRXChannel(Short channel) {
 		this.channel = channel;
 	}
 
-	public byte getChannel() {
+	public Short getChannel() {
 		return channel;
 	}
 
-	public void setChannel(byte channel) {
+	public void setChannel(Short channel) {
 		this.channel = channel;
 	}
